+++ src/weak_key_hash_map.rs
@@ -803,6 +803,8 @@ impl<'a, K: WeakElement, V, F> Iterator for ExtractIf<'a, K, V, F> {
 #[cfg(test)]
 mod test {
     #![allow(clippy::print_stderr)]
+    // TODO 050: remove.
+    #![cfg_attr(feature = "ahash", allow(deprecated))]
 
     use super::{Entry, WeakKeyHashMap};
     use crate::{