+++ src/weak_value_hash_map.rs
@@ -603,6 +603,9 @@ impl<'a, K, V: WeakElement, F> Iterator for ExtractIf<'a, K, V, F> {
 
 #[cfg(test)]
 mod test {
+    // TODO 050: remove.
+    #![cfg_attr(feature = "ahash", allow(deprecated))]
+
     use super::WeakValueHashMap;
     use crate::{
         compat::{